Move window procedure in submodule
