+++ src/hooks/common/wnd_proc.rs
@@ -3,22 +3,25 @@
 use std::mem::size_of;
 
 use imgui::Io;
-use windows::Win32::Foundation::{LPARAM, WPARAM};
+use parking_lot::MutexGuard;
+use windows::Win32::Foundation::{HWND, LPARAM, LRESULT, WPARAM};
 use windows::Win32::UI::Input::KeyboardAndMouse::{
-    MapVirtualKeyA, VIRTUAL_KEY, VK_CONTROL, VK_LBUTTON, VK_MBUTTON, VK_MENU, VK_RBUTTON, VK_SHIFT,
+    MapVirtualKeyA, VIRTUAL_KEY, VK_CONTROL, VK_LBUTTON, VK_LCONTROL, VK_LMENU, VK_LSHIFT, VK_LWIN,
+    VK_MBUTTON, VK_MENU, VK_RBUTTON, VK_RCONTROL, VK_RMENU, VK_RSHIFT, VK_RWIN, VK_SHIFT,
     VK_XBUTTON1, VK_XBUTTON2,
 };
 use windows::Win32::UI::Input::{
     GetRawInputData, HRAWINPUT, RAWINPUT, RAWINPUTHEADER, RAWKEYBOARD, RAWMOUSE_0_0,
     RID_DEVICE_INFO_TYPE, RID_INPUT, RIM_TYPEKEYBOARD, RIM_TYPEMOUSE,
 };
-use windows::Win32::UI::WindowsAndMessaging::{
-    MAPVK_VSC_TO_VK_EX, RIM_INPUT, RI_KEY_BREAK, RI_KEY_E0, RI_KEY_E1, RI_KEY_MAKE,
-    RI_MOUSE_BUTTON_4_DOWN, RI_MOUSE_BUTTON_4_UP, RI_MOUSE_BUTTON_5_DOWN, RI_MOUSE_BUTTON_5_UP,
-    RI_MOUSE_HWHEEL, RI_MOUSE_LEFT_BUTTON_DOWN, RI_MOUSE_LEFT_BUTTON_UP,
-    RI_MOUSE_MIDDLE_BUTTON_DOWN, RI_MOUSE_MIDDLE_BUTTON_UP, RI_MOUSE_RIGHT_BUTTON_DOWN,
-    RI_MOUSE_RIGHT_BUTTON_UP, RI_MOUSE_WHEEL, WHEEL_DELTA,
-};
+use windows::Win32::UI::WindowsAndMessaging::*;
+
+use super::{ImguiRenderLoop, ImguiWindowsEventHandler};
+use crate::hooks::{get_wheel_delta_wparam, hiword, loword};
+
+////////////////////////////////////////////////////////////////////////////////
+// Raw input
+////////////////////////////////////////////////////////////////////////////////
 
 // Handle raw mouse input events.
 //
@@ -137,7 +140,7 @@ fn handle_raw_keyboard_input(io: &mut Io, raw_keyboard: &RAWKEYBOARD) {
 }
 
 // Handle WM_INPUT events.
-pub(crate) fn handle_raw_input(io: &mut Io, WPARAM(wparam): WPARAM, LPARAM(lparam): LPARAM) {
+fn handle_raw_input(io: &mut Io, WPARAM(wparam): WPARAM, LPARAM(lparam): LPARAM) {
     let mut raw_data = RAWINPUT { ..Default::default() };
     let mut raw_data_size = size_of::<RAWINPUT>() as u32;
     let raw_data_header_size = size_of::<RAWINPUTHEADER>() as u32;
@@ -173,4 +176,131 @@ pub(crate) fn handle_raw_input(io: &mut Io, WPARAM(wparam): WPARAM, LPARAM(lpara
         },
         _ => {},
     }
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Regular input
+////////////////////////////////////////////////////////////////////////////////
+
+fn map_vkey(wparam: u16, lparam: usize) -> VIRTUAL_KEY {
+    match VIRTUAL_KEY(wparam) {
+        VK_SHIFT => unsafe {
+            match MapVirtualKeyA(((lparam & 0x00ff0000) >> 16) as u32, MAPVK_VSC_TO_VK_EX) {
+                0 => VIRTUAL_KEY(wparam),
+                i => VIRTUAL_KEY(i as _),
+            }
+        },
+        VK_CONTROL => {
+            if lparam & 0x01000000 != 0 {
+                VK_RCONTROL
+            } else {
+                VK_LCONTROL
+            }
+        },
+        VK_MENU => {
+            if lparam & 0x01000000 != 0 {
+                VK_RMENU
+            } else {
+                VK_LMENU
+            }
+        },
+        _ => VIRTUAL_KEY(wparam),
+    }
+}
+
+// Handle WM_(SYS)KEYDOWN/WM_(SYS)KEYUP events.
+fn handle_input(io: &mut Io, state: u32, WPARAM(wparam): WPARAM, LPARAM(lparam): LPARAM) {
+    let pressed = (state == WM_KEYDOWN) || (state == WM_SYSKEYDOWN);
+    let key_pressed = map_vkey(wparam as _, lparam as _);
+    io.keys_down[key_pressed.0 as usize] = pressed;
+
+    // According to the winit implementation [1], it's ok to check twice, and the
+    // logic isn't flawed either.
+    //
+    // [1] https://github.com/imgui-rs/imgui-rs/blob/b1e66d050e84dbb2120001d16ce59d15ef6b5303/imgui-winit-support/src/lib.rs#L401-L404
+    match key_pressed {
+        VK_CONTROL | VK_LCONTROL | VK_RCONTROL => io.key_ctrl = pressed,
+        VK_SHIFT | VK_LSHIFT | VK_RSHIFT => io.key_shift = pressed,
+        VK_MENU | VK_LMENU | VK_RMENU => io.key_alt = pressed,
+        VK_LWIN | VK_RWIN => io.key_super = pressed,
+        _ => (),
+    };
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Window procedure
+////////////////////////////////////////////////////////////////////////////////
+
+#[must_use]
+pub(crate) fn imgui_wnd_proc_impl<T>(
+    hwnd: HWND,
+    umsg: u32,
+    WPARAM(wparam): WPARAM,
+    LPARAM(lparam): LPARAM,
+    mut imgui_renderer: MutexGuard<Box<impl ImguiWindowsEventHandler>>,
+    imgui_render_loop: T,
+) -> LRESULT
+where
+    T: AsRef<dyn Send + Sync + ImguiRenderLoop + 'static>,
+{
+    let io = imgui_renderer.io_mut();
+    match umsg {
+        WM_INPUT => handle_raw_input(io, WPARAM(wparam), LPARAM(lparam)),
+        state @ (WM_KEYDOWN | WM_SYSKEYDOWN | WM_KEYUP | WM_SYSKEYUP) if wparam < 256 => {
+            handle_input(io, state, WPARAM(wparam), LPARAM(lparam))
+        },
+        WM_LBUTTONDOWN | WM_LBUTTONDBLCLK => {
+            io.mouse_down[0] = true;
+        },
+        WM_RBUTTONDOWN | WM_RBUTTONDBLCLK => {
+            io.mouse_down[1] = true;
+        },
+        WM_MBUTTONDOWN | WM_MBUTTONDBLCLK => {
+            io.mouse_down[2] = true;
+        },
+        WM_XBUTTONDOWN | WM_XBUTTONDBLCLK => {
+            let btn = if hiword(wparam as _) == XBUTTON1.0 as u16 { 3 } else { 4 };
+            io.mouse_down[btn] = true;
+        },
+        WM_LBUTTONUP => {
+            io.mouse_down[0] = false;
+        },
+        WM_RBUTTONUP => {
+            io.mouse_down[1] = false;
+        },
+        WM_MBUTTONUP => {
+            io.mouse_down[2] = false;
+        },
+        WM_XBUTTONUP => {
+            let btn = if hiword(wparam as _) == XBUTTON1.0 as u16 { 3 } else { 4 };
+            io.mouse_down[btn] = false;
+        },
+        WM_MOUSEWHEEL => {
+            let wheel_delta_wparam = get_wheel_delta_wparam(wparam as _);
+            let wheel_delta = WHEEL_DELTA as f32;
+            io.mouse_wheel += (wheel_delta_wparam as i16 as f32) / wheel_delta;
+        },
+        WM_MOUSEHWHEEL => {
+            let wheel_delta_wparam = get_wheel_delta_wparam(wparam as _);
+            let wheel_delta = WHEEL_DELTA as f32;
+            io.mouse_wheel_h += (wheel_delta_wparam as i16 as f32) / wheel_delta;
+        },
+        WM_CHAR => io.add_input_character(wparam as u8 as char),
+        WM_ACTIVATE => {
+            *imgui_renderer.focus_mut() = loword(wparam as _) != WA_INACTIVE as u16;
+            return LRESULT(1);
+        },
+        _ => {},
+    };
+
+    let wnd_proc = imgui_renderer.wnd_proc();
+    let should_block_messages =
+        imgui_render_loop.as_ref().should_block_messages(imgui_renderer.io());
+    drop(imgui_renderer);
+
+    if should_block_messages {
+        return LRESULT(1);
+    }
+
+    unsafe { CallWindowProcW(Some(wnd_proc), hwnd, umsg, WPARAM(wparam), LPARAM(lparam)) }
 }
